auth middleware

+++ backend/controllers/goalControllers.js
@@ -1,8 +1,9 @@
 import asyncHandler from "express-async-handler";
 import goalsModel from "../models/goalModel.js";
+import userModel from "../models/userModel.js";
 
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await goalsModel.find();
+  const goals = await goalsModel.find({ user: req.user.id });
   if (!goals) {
     res.status(404);
     throw new Error("No goals found here");
@@ -17,6 +18,7 @@ const setGoals = asyncHandler(async (req, res) => {
   }
   const newGoal = await goalsModel({
     text: req.body.text,
+    user: req.user.id,
   });
   await newGoal.save();
   res.status(200).json(newGoal);
@@ -28,6 +30,19 @@ const updateGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Could not find goal");
   }
+
+  const user = await userModel.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("Unauthorized, bad token");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Unauthorized, bad token");
+  }
+
   const updatedGoal = await goalsModel.findOneAndUpdate(
     { _id: req.params.id },
     req.body,
@@ -42,6 +57,18 @@ const deleteGoals = asyncHandler(async (req, res) => {
     res.status(400);
     throw new Error("Could not find goal");
   }
+  const user = await userModel.findById(req.user.id);
+
+  if (!user) {
+    res.status(401);
+    throw new Error("Unauthorized, bad token");
+  }
+
+  if (goal.user.toString() !== user.id) {
+    res.status(401);
+    throw new Error("Unauthorized, bad token");
+  }
+
   await goal.deleteOne();
   res.status(200).json({ id: req.params.id });
 });
